+++ src/Git_Updater/API/GitHub_API.php
@@ -171,8 +171,6 @@ class GitHub_API extends API implements API_Interface {
 		}
 
 		$download_link = $download_link_base . $endpoint;
-
-		$download_link = apply_filters_deprecated( 'github_updater_post_construct_download_link', [ $download_link, $this->type, $branch_switch ], '10.0.0', 'gu_post_construct_download_link' );
 		$download_link = apply_filters( 'gu_post_construct_download_link', $download_link, $this->type, $branch_switch );
 
 		return $download_link;
@@ -286,6 +284,7 @@ class GitHub_API extends API implements API_Interface {
 			function ( $e ) use ( &$arr ) {
 				$arr['private']      = $e->private;
 				$arr['last_updated'] = $e->pushed_at;
+				$arr['added']        = $e->created_at;
 				$arr['watchers']     = $e->watchers;
 				$arr['forks']        = $e->forks;
 				$arr['open_issues']  = $e->open_issues;
@@ -397,10 +396,10 @@ class GitHub_API extends API implements API_Interface {
 		$files = [];
 		$dirs  = [];
 		foreach ( $response as $content ) {
-			if ( 'file' === $content->type ) {
+			if ( property_exists( $content, 'type' ) && 'file' === $content->type ) {
 				$files[] = $content->name;
 			}
-			if ( 'dir' === $content->type ) {
+			if ( property_exists( $content, 'type' ) && 'dir' === $content->type ) {
 				$dirs[] = $content->name;
 			}
 		}
@@ -547,7 +546,7 @@ class GitHub_API extends API implements API_Interface {
 			<input class="github_setting" type="password" style="width:50%;" id="github_access_token" name="github_access_token" value="" autocomplete="new-password">
 			<br>
 			<span class="description">
-				<?php esc_html_e( 'Enter GitHub Access Token for private GitHub repositories.', 'git-updater' ); ?>
+			<?php esc_html_e( 'Enter GitHub Access Token for private GitHub repositories.', 'git-updater' ); ?>
 			</span>
 		</label>
 		<?php