+++ src/Git_Updater/Plugin.php
@@ -133,9 +133,7 @@ class Plugin {
 		);
 
 		$additions = apply_filters( 'gu_additions', null, $plugins, 'plugin' );
-		$additions = null === $additions ? apply_filters_deprecated( 'github_updater_additions', [ null, $plugins, 'plugin' ], '10.0.0', 'gu_additions' ) : $additions;
-
-		$plugins = array_merge( $plugins, (array) $additions );
+		$plugins   = array_merge( $plugins, (array) $additions );
 		ksort( $plugins );
 
 		foreach ( (array) $plugins as $slug => $plugin ) {
@@ -175,6 +173,7 @@ class Plugin {
 
 			$git_plugin['type']           = 'plugin';
 			$git_plugin['git']            = $repo_parts['git_server'];
+			$git_plugin['did']            = $header['did'];
 			$git_plugin['uri']            = "{$header['base_uri']}/{$header['owner_repo']}";
 			$git_plugin['enterprise']     = $header['enterprise_uri'];
 			$git_plugin['enterprise_api'] = $header['enterprise_api'];
@@ -239,7 +238,6 @@ class Plugin {
 		$config = apply_filters( 'gu_config_pre_process', $this->config );
 
 		$disable_wp_cron = (bool) apply_filters( 'gu_disable_wpcron', false );
-		$disable_wp_cron = $disable_wp_cron ?: (bool) apply_filters_deprecated( 'github_updater_disable_wpcron', [ false ], '10.0.0', 'gu_disable_wpcron' );
 
 		foreach ( (array) $config as $plugin ) {
 			if ( ! $this->waiting_for_background_update( $plugin ) || static::is_wp_cli() || $disable_wp_cron ) {
@@ -303,6 +301,7 @@ class Plugin {
 			return $result;
 		}
 
+		$response->did         = $plugin->did;
 		$response->slug        = $plugin->slug;
 		$response->plugin_name = $plugin->name;
 		$response->name        = $plugin->name;
@@ -318,8 +317,9 @@ class Plugin {
 		$response->tested            = $plugin->tested;
 		$response->downloaded        = $plugin->downloaded ?: 0;
 		$response->active_installs   = $response->downloaded;
-		$response->last_updated      = $plugin->last_updated ?: null;
-		$response->download_link     = $plugin->download_link ?: null;
+		$response->last_updated      = $plugin->last_updated ?: '';
+		$response->added             = $plugin->added ?: '';
+		$response->download_link     = $plugin->download_link ?: '';
 		$response->banners           = $plugin->banners;
 		$response->icons             = $plugin->icons ?: [];
 		$response->contributors      = $plugin->contributors;
@@ -408,7 +408,6 @@ class Plugin {
 					}
 
 					$overrides = apply_filters( 'gu_override_dot_org', [] );
-					$overrides = empty( $overrides ) ? apply_filters_deprecated( 'github_updater_override_dot_org', [ [] ], '10.0.0', 'gu_override_dot_org' ) : $overrides;
 
 					if ( isset( $transient->response[ $plugin->file ] ) && in_array( $plugin->file, $overrides, true ) ) {
 						unset( $transient->response[ $plugin->file ] );