+++ src/Git_Updater/REST/REST_API.php
@@ -450,6 +450,7 @@ class REST_API {
 		}
 
 		$repo_api_data = [
+			'did'               => $repo_data->did,
 			'name'              => $repo_data->name,
 			'slug'              => $repo_data->slug,
 			'git'               => $repo_data->git,
@@ -476,7 +477,8 @@ class REST_API {
 			'donate_link'       => $repo_data->donate_link,
 			'banners'           => $repo_data->banners,
 			'icons'             => $repo_data->icons,
-			'last_updated'      => $repo_data->last_updated,
+			'last_updated'      => gmdate( 'Y-m-d h:ia T', strtotime( $repo_data->last_updated ) ),
+			'added'             => gmdate( 'Y-m-d', strtotime( $repo_data->added ) ),
 			'num_ratings'       => $repo_data->num_ratings,
 			'rating'            => $repo_data->rating,
 			'active_installs'   => $repo_data->downloaded,