+++ src/Git_Updater/Theme.php
@@ -149,9 +149,7 @@ class Theme {
 		);
 
 		$additions = apply_filters( 'gu_additions', null, $themes, 'theme' );
-		$additions = null === $additions ? apply_filters_deprecated( 'github_updater_additions', [ null, $themes, 'theme' ], '10.0.0', 'gu_additions' ) : $additions;
-
-		$themes = array_merge( $themes, (array) $additions );
+		$themes    = array_merge( $themes, (array) $additions );
 		ksort( $themes );
 
 		foreach ( (array) $themes as $slug => $theme ) {
@@ -191,6 +189,7 @@ class Theme {
 
 			$git_theme['type']           = 'theme';
 			$git_theme['git']            = $repo_parts['git_server'];
+			$git_theme['did']            = $header['did'];
 			$git_theme['uri']            = "{$header['base_uri']}/{$header['owner_repo']}";
 			$git_theme['theme_uri']      = $header['owner_repo'];
 			$git_theme['enterprise']     = $header['enterprise_uri'];
@@ -245,7 +244,6 @@ class Theme {
 		$config = apply_filters( 'gu_config_pre_process', $this->config );
 
 		$disable_wp_cron = (bool) apply_filters( 'gu_disable_wpcron', false );
-		$disable_wp_cron = $disable_wp_cron ?: (bool) apply_filters_deprecated( 'github_updater_disable_wpcron', [ false ], '10.0.0', 'gu_disable_wpcron' );
 
 		foreach ( (array) $config as $theme ) {
 			if ( ! $this->waiting_for_background_update( $theme ) || static::is_wp_cli() || $disable_wp_cron
@@ -317,6 +315,7 @@ class Theme {
 			return $result;
 		}
 
+		$response->did          = $theme->did;
 		$response->slug         = $theme->slug;
 		$response->name         = $theme->name;
 		$response->homepage     = $theme->homepage;
@@ -634,7 +633,6 @@ class Theme {
 				}
 
 				$overrides = apply_filters( 'gu_override_dot_org', [] );
-				$overrides = empty( $overrides ) ? apply_filters_deprecated( 'github_updater_override_dot_org', [ [] ], '10.0.0', 'gu_override_dot_org' ) : $overrides;
 
 				if ( isset( $transient->response[ $theme->slug ] ) && in_array( $theme->slug, $overrides, true ) ) {
 					unset( $transient->response[ $theme->slug ] );