+++ CHANGES.md
@@ -1,5 +1,14 @@
 #### [unreleased]
 
+#### 12.15.0 / 2025-05-20
+* update to correct format of readme tags
+* add correctly formated date/time for `update-api` REST endpoint
+* remove deprecated hooks from v10 and earlier
+* add error checking to `parse_contents_response()`
+* update Freemius/wordpress-sdk
+* add support for `Plugin ID` and `Theme ID` headers for FAIR
+* update `composer.json`
+
 #### 12.14.0 / 2025-02-26
 * make sure proper release asset headers are added even if access token not set
 * ensure _short description_ is 150 characters or less